Initial re-organization.
Re-organized the content and implemented a new theme/organization concept. This is not really usable yet, but it is an important step to fitting more information into the docs in a more accessible way.

For this we are using a Bootstrap theme, plus custom enhancements to SyntaxHighlighter, including a custom Lex highlighter that does some custom highlighting.